requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
drf-yasg>=1.20.0
django-cors-headers>=3.7.0
python-dotenv>=0.19.0
djangorestframework-simplejwt>=5.4.0